+++ decidim-initiatives/spec/system/admin/update_initiative_spec.rb
@@ -35,6 +35,12 @@ describe "User prints the initiative", type: :system do
           end
           submit_and_validate
         end
+
+        it "displays initiative attachments" do
+          page.find(".action-icon--edit").click
+          expect(page).to have_link("Edit")
+          expect(page).to have_link("New")
+        end
       end
 
       context "when initiative is in validating state" do
@@ -51,6 +57,12 @@ describe "User prints the initiative", type: :system do
             expect(page).to have_css("#initiative_signature_type[disabled]")
           end
           expect(page).to have_no_css("*[type=submit]")
+        end
+
+        it "displays initiative attachments" do
+          page.find(".action-icon--edit").click
+          expect(page).to have_link("Edit", class: "disabled")
+          expect(page).to have_link("New", class: "disabled")
         end
       end
     end
@@ -84,6 +96,12 @@ describe "User prints the initiative", type: :system do
           end
           submit_and_validate
         end
+
+        it "displays initiative attachments" do
+          page.find(".action-icon--edit").click
+          expect(page).to have_link("Edit")
+          expect(page).to have_link("New")
+        end
       end
 
       context "when initiative is in validating state" do
@@ -100,6 +118,12 @@ describe "User prints the initiative", type: :system do
           end
           submit_and_validate
         end
+
+        it "displays initiative attachments" do
+          page.find(".action-icon--edit").click
+          expect(page).to have_link("Edit")
+          expect(page).to have_link("New")
+        end
       end
 
       context "when initiative is in accepted state" do
@@ -115,6 +139,12 @@ describe "User prints the initiative", type: :system do
             expect(page).to have_css("#initiative_decidim_scope_id[disabled]")
             expect(page).to have_css("#initiative_signature_type[disabled]")
           end
+        end
+
+        it "displays initiative attachments" do
+          page.find(".action-icon--edit").click
+          expect(page).to have_link("Edit")
+          expect(page).to have_link("New")
         end
       end
     end